Only 1 bear instance attacks when detect player. "Edgar Playground" scene renamed to "Howl PS demo".
Event system was activating on all bear instances before (due to the logic of the code as well as the event variable being static, which activates for all instances of script/game object). This was replaced with code that references the script from the parent so that the child (which detects if wolf is detected by bear) can call the method ONLY from it's parent's script, not from all instances of the script.