Game Over Fix + Knock Down Trees + Hurt Flash
-  Game Over in a tutorial section would send player to the original
main area scene. Fixed this to reset it to the current scene.

- If the player hits the pushable trees from the left or right side, it
will now fall over in the correct direction.

- Sort of fixed the hurt flash. When hurt, it'll gradually increase its
red color, then go back to the original color. The wolf is also
invincible during this animation. (This still needs the coloring to be
fixed + maybe have a seperate, longer invincible timer.)