Changes for focus on theme of Home/Comfort/ Coziness:
-Added an "Affection" animation
For now, the "V" key activates this, which makes your wolf "cuddle". If close to another wolf, they show affection too.
This isn't present in all scenes yet and only works when wolf isn't following you.

-Run meter changed to "Warmth Meter"
The cold decreases your wolf's body heat. To regain it, you have to show affection/cuddle up to another wolf.
It's currently active on the player (adult wolf) for testing but intended for wolf pups (as they can't keep body heat the way adults can) and wolves with mange (loss of fur due to parasites. This depends on whether or not we include this in the game later for story/mechanics)
When warmth meter is empty, wolf is too cold to run. This emphasizes closeness. Player will want to run and they have the choice on regaining their ability to run.

-New Den scene