Added a 'cold vignetting' system - as discussed.
Basically the player now has a "Warmth Temperature" value that, as it lowers, vignettes the screen slowly.  If the player collides with an object that has a "Warmth Object" script, it'll refill that value until it hits 50, the max value. There's nothing that happens when the value hits 0, but the current plan is to switch the player from the Spirit World to the Real World when this happens.

A scene with this working as an example can be found named "Vignetting Tests".